requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.2.0
backcall==0.2.0
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175564434/work
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.3.2
cycler==0.11.0
decorator==5.1.1
deepchem==2.7.1
dgl==1.1.3
fair-esm==2.0.0
filelock==3.12.2
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
gensim==4.2.0
h5py==3.8.0
huggingface-hub==0.16.4
idna==3.6
importlib-metadata==6.7.0
ipython==7.34.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
lightning-utilities==0.10.1
littleutils==0.2.2
MarkupSafe==2.1.4
matplotlib==3.5.3
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
mol2vec==0.2.2
multidict==6.0.4
networkx==2.6.3
numexpr @ file:///croot/numexpr_1668713893690/work
numpy==1.21.6
ogb==1.3.5
omegaconf==2.3.0
outdated==0.2.2
packaging==23.2
pandas==1.3.5
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytorch-lightning==1.9.5
pytz==2023.3.post1
PyYAML==6.0.1
rdkit==2023.3.2
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
sentencepiece==0.2.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==6.4.0
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==1.13.0+cu117
torch-cluster @ file:///home/inspur/zdp409100230054/torch_cluster-1.6.0%2Bpt113cu117-cp37-cp37m-linux_x86_64.whl
torch-geometric==2.3.1
torch-scatter @ file:///home/inspur/zdp409100230054/torch_scatter-2.1.0%2Bpt113cu117-cp37-cp37m-linux_x86_64.whl
torch-sparse @ file:///home/inspur/zdp409100230054/torch_sparse-0.6.15%2Bpt113cu117-cp37-cp37m-linux_x86_64.whl
torch-spline-conv @ file:///home/inspur/zdp409100230054/torch_spline_conv-1.2.1%2Bpt113cu117-cp37-cp37m-linux_x86_64.whl
torchaudio==0.13.0+cu117
torchmetrics==0.11.4
torchvision==0.14.0+cu117
tqdm==4.66.1
traitlets==5.9.0
transformers==4.30.2
typing_extensions==4.7.1
urllib3==2.0.7
wcwidth==0.2.13
yarl==1.9.4
zipp==3.15.0
